refactor : 해결되지 않은 경로 에러 해결
