+++ src/components/user/mypage/ContentTab.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react';
 import * as S from './ContentTab.styled';
 import { Link, Outlet, useLocation } from 'react-router-dom';
-import { ROUTES } from '../../constants/routes';
 import ScrollWrapper from './ScrollWrapper';
 import MovedInquiredLink from '../customerService/MoveInquiredLink';
+import { ROUTES } from '../../../constants/user/routes';
 
 interface Filter {
   title: string;
