+++ src/constants/sidebarItems.tsx
@@ -1,9 +1,9 @@
-import { ROUTES } from './routes';
 import {
   UserGroupIcon,
   PencilSquareIcon,
   UserPlusIcon,
 } from '@heroicons/react/24/outline';
+import { ROUTES } from './user/routes';
 
 export const applicantsMenuItems = (projectId: number, isDone?: boolean) => {
   return [
